+++ base/src/Data/Macaw/CFG.hs
@@ -9,10 +9,8 @@ module Data.Macaw.CFG
   , module Data.Macaw.CFG.App
   , module Data.Macaw.Memory
   , Data.Macaw.Types.FloatInfoRepr(..)
-  , Data.Macaw.Architecture.Info.ArchBlockPrecond
   ) where
 
-import qualified Data.Macaw.Architecture.Info
 import           Data.Macaw.CFG.App
 import           Data.Macaw.CFG.Core
 import           Data.Macaw.Memory
