+++ refinement/tools/Summary.hs
@@ -14,8 +14,8 @@ import           Data.Monoid
 import           Data.Parameterized.Some ( Some(..) )
 import           Data.Semigroup
 import qualified Data.Set as S
-import           Data.Text.Prettyprint.Doc as PP
 import           Data.Word ( Word64 )
+import           Prettyprinter as PP
 
 import qualified Data.Macaw.CFG as MC
 import qualified Data.Macaw.Discovery as MD
